pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["daniskzn <danis.sharafiev@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
ruff = "^0.3.2"
drf-spectacular = "^0.27.1"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.1"
requests = "^2.31.0"
django-debug-toolbar = "^4.3.0"
pika = "^1.3.2"
faststream = {extras = ["rabbit"], version = "^0.4.7"}
uvicorn = {extras = ["standard"], version = "^0.28.0"}
django-split-settings = "^1.3.0"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.1"
sqlalchemy = "^2.0.29"
django-tenants = "^3.6.1"
pyjwt = "^2.8.0"
python-telegram-bot = "^21.1"
cryptography = "^42.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
